WIP: Migrate to the Heimdall2 backend
This commit takes a lot of the logic that has been copy-pasted between different files and moves it all to a global location.
Initial page load is now paused until we figure out what mode we are in (since we have vue-router wait until we know before it proceeds). We check whether or not `/server` responds with 200 or 404, if it responds with 200 then we know we are in server mode and proceed accordingly.

Before this is merged, the backend.ts Vuex store should be renamed to server.ts and the current server.ts should be removed (once all the functionality has been ported). backend.ts is intended to eventually be a full-featured replacement for server.ts.