Modify declaration of Vuex store and reuse the declaration everwhere.
This gets rid of all calls to getModule in the application and instead places the const declaration in their appropriate store files.

Run all SVG assets through https://jakearchibald.github.io/svgomg/. This is suggested by https://vue-svg-loader.js.org/faq.html#how-to-use-this-loader-with-jest which is the plugin we use for loading SVGs, and fixes issues we will encounter in the test suite when trying to load components that use SVGs.